+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.3.0
+* Handle multiple inputs
+* Resize advanced window when images are displayed
+* Add eye laterality classification
+* Enable parallel step execution
+* Add optic disc segmentation
+* Clean arterial signal (used for artery/vein segmentation): remove heartbeats badly correlated with median beat
+
 ## 1.2.0
 * Take into account *NumberOfWorkers* params. By default, set at 0.5 : take half of the available workers
 * Enable loading of doppler_view config